+++ PaymentSystem/IPaymentProcessor.cs
@@ -2,5 +2,5 @@ namespace PaymentSystem;
 
 public interface IPaymentProcessor
 {
-    PaymentStatus ProcessPayment(double amount);
+    PaymentStatus ProcessPayment(decimal amount);
 }
